packages/notte-core/pyproject.toml
[project]
name = "notte-core"
version = "1.3.3"
description = "The web browser for LLMs agents"
readme = "README.md"
authors = [
    { name = "Notte Team ", email = "hello@notte.cc" }
]
packages = [
  { include = "notte_core", from = "src" },
]

requires-python = ">=3.11"
dependencies = [
    "chevron>=0.14.0",
    "litellm>=1.61.16",
    "llamux>=0.1.9",
    "loguru>=0.7.3",
    "patchright==1.50.0",
    "pillow>=11.1.0",
    "posthog>=3.0.1",
    "pydantic>=2.10.6",
    "pyotp>=2.9.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "tldextract>=5.2.0",
]

[project.optional-dependencies]
server = [
    "litellm[proxy]>=1.61.16",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
maincontentextractor = { git = "https://github.com/HawkClaws/main_content_extractor", rev = "7c3ed7f6ed7f6c10223a3357d43ab741663bc812" }
